% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_calc.R
\name{dd_case}
\alias{dd_case}
\title{Determine the case
Determine the relationship between the daily min and max and the upper & lower thresholds}
\usage{
dd_case(daily_min, daily_max, thresh_low, thresh_up)
}
\description{
Determine the case
Determine the relationship between the daily min and max and the upper & lower thresholds
}
\keyword{internal}
