% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_calc.R
\name{dd_dblsine_half}
\alias{dd_dblsine_half}
\title{Compute half-day GDD using sine functions.
This is an internal function, called by dd_calc()}
\usage{
dd_dblsine_half(case, tmin, tmax, thresh_low, thresh_up)
}
\description{
Compute half-day GDD using sine functions.
This is an internal function, called by dd_calc()
}
\keyword{internal}
