% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fetch_cimis.R
\name{dd_fetch_cimis}
\alias{dd_fetch_cimis}
\title{Fetch daily temperature data from CIMIS}
\usage{
dd_fetch_cimis(
  loc,
  start_date,
  end_date,
  api_key,
  use_units = FALSE,
  debug = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{loc}{The location}

\item{start_date}{The start date}

\item{end_date}{The end date}

\item{api_key}{An API key}

\item{use_units}{Return temperature data as a units object}

\item{debug}{Show additional messages}

\item{quiet}{Suppress messages}
}
\value{
A tibble with the columns date, station, tmin, and tmax. tmin and tmax will be in Farenheit.
}
\description{
Fetch daily temperature data from CIMIS
}
\details{
\code{loc} can be a numeric vector containing longitude and latitude coordinates, or the numeric id
of a CIMIS station.
}
