% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_models.R
\docType{data}
\name{dd_models}
\alias{dd_models}
\title{Degree day models}
\format{
A tibble with the lower and upper threshold and methods for modeling development
}
\source{
\url{http://ipm.ucanr.edu/WEATHER/DDTABLE/index.html}
}
\usage{
data(dd_models)
}
\description{
Models of degree days for various insects and plants
}
\details{
The following climate variables are available as raster series thru the Cal-Adapt API.
}
\keyword{datasets}
