% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_calc.R
\name{dd_calc}
\alias{dd_calc}
\alias{dd_sng_tri}
\alias{dd_sng_sine}
\alias{dd_dbl_tri}
\alias{dd_dbl_sine}
\title{Compute degree days from daily data}
\usage{
dd_calc(
  daily_min,
  daily_max,
  nextday_min = NULL,
  thresh_low = NULL,
  thresh_up = NULL,
  method = c("sng_tri", "dbl_tri", "sng_sine", "dbl_sine", "simp_avg")[0],
  cutoff = c("horizontal", "vertical")[1],
  cumulative = FALSE,
  no_neg = TRUE,
  interpolate_na = FALSE,
  quiet = FALSE,
  debug = FALSE
)

dd_sng_tri(
  daily_min,
  daily_max,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

dd_sng_sine(
  daily_min,
  daily_max,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

dd_dbl_tri(
  daily_min,
  daily_max,
  nextday_min = NULL,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

dd_dbl_sine(
  daily_min,
  daily_max,
  nextday_min = NULL,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)
}
\arguments{
\item{daily_min}{The daily minimum temperature}

\item{daily_max}{The daily maximum temperature}

\item{nextday_min}{The minimum temp the day after}

\item{thresh_low}{The lower development threshold}

\item{thresh_up}{The upper development threshold}

\item{method}{The method used}

\item{cutoff}{The cutoff method}

\item{cumulative}{Return cumulative values}

\item{no_neg}{Set negative values to zero}

\item{interpolate_na}{Interpolate missing values}

\item{quiet}{Suppress messages}

\item{debug}{Show additional messages}
}
\description{
Compute degree days from daily min and max temperature
}
\details{
Units for \code{daily_min}, \code{daily_max}, \code{thresh_low}, and \code{thresh_up} should all be the same
(i.e., all Fahrenheit or all Celsius). The function does not check for unit consistency.

\code{nextday_min} is required for the double-triangle and the double-sine methods. These methods use the minimum temperature
of the following day to model temperatures in the 2nd half of thd day.

\code{no_neg = TRUE} sets negative values to zero. This is generally preferred when using degree days to predict the timing of
 development milestones, if one assumes that growth can not go backwards.

Missing values (NAs) in the temperatures will result in NA degree days. If \code{interpolate_na = TRUE}, missing degree days will
be interpolated. NAs in the middle of the series will be linearly interpolated, and NAs at the ends will be filled with the
adjacent values.
}
\section{Functions}{
\itemize{
\item \code{dd_sng_tri()}: Compute degree days using the single-triangle method

\item \code{dd_sng_sine()}: Compute degree days using the single-sine method

\item \code{dd_dbl_tri()}: Compute degree days using the double-triangle method

\item \code{dd_dbl_sine()}: Compute degree days using the double-sine method

}}
